requirements.txt
pandas>=2.1
pyarrow>=12.0
sacrebleu>=2.4
unbabel-comet>=2.2
torch>=2.2
transformers>=4.38
sentencepiece>=0.1.99
tqdm>=4.65
matplotlib>=3.7

